+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
@@ -6,18 +7,33 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-## 0.9.4 - 2024-01-29
 ### Added
+
+- Add `LazySeq.reduce` to match the API of the
+  [TC39 proposal](https://github.com/tc39/proposal-iterator-helpers)
+
+### Changed
+
+- **Breaking** Renamed `LazySeq.anyMatch` to `LazySeq.some` and renamed `LazySeq.allMatch` to `LazySeq.every`
+  to match the API of the [TC39 proposal](https://github.com/tc39/proposal-iterator-helpers)
+
+## 0.9.4 - 2024-01-29
+
+### Added
+
 - Lots of documentation improvements.
 - A few performance improvements.
 - Add LazySeq.toOrderedSet to convert a sequence to an ordered set.
 - Add `transform` function to HashMap, HashSet, OrderedMap, and OrderedSet.
 
 ### Changed
+
 - **Breaking** Rename `LazySeq.foldLeft` to just `LazySeq.fold`.
 
 ## 0.9.3 - 2022-08-03
+
 ### Added
+
 - Add LazySeq.distinctAndSortBy which allows distinct and also sorting by one or more
   properties.
 - Add deleteMin, deleteMax, lookupMin, lookupMax to OrderedMap
@@ -25,6 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add OrderedSet class
 
 ### Changed
+
 - **Breaking**: Rename `LazySeq.ofIterable` to just `LazySeq.of`. This is used frequently and a short name
   is better, and this is the last chance before the 1.0 API release.
 - The LazySeq.filter function now can restrict the type if the function has a type guard.
@@ -32,16 +49,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   once.
 
 ## 0.9.2 - 2022-07-21
+
 ### Changed
+
 - The toLazySeq (and friends) methods on HashMap, HashSet, and OrderedMap now correctly
   allow iteration more than once (so implementing the iterable protocol).
 
 ## 0.9.1 - 2022-07-15
+
 ### Changed
+
 - rename LazySeq maxOn, minOn, sort, and sumOn to maxBy, minBy, sortBy, and sumBy
   to be consistent with the rest of the methods on LazySeq
 - require at least one argument to LazySeq.distinctBy, LazySeq.maxBy, LazySeq.minBy,
   LazySeq.sortBy LazySeq.toSortedArray
 
 ## 0.9.0 - 2022-07-13
+
 Initial release, implementing `HashMap`, `HashSet`, `OrderedMap`, and `LazySeq`.